requirements.txt
glcontext==3.0.0
glfw==2.8.0
numpy==2.2.3
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
